Function fedimint_testing::db::validate_migrations_global

pub async fn validate_migrations_global<F, Fut>(
    validate: F,
    db_prefix: &str,
    target_db_version: DatabaseVersion,
    migrations: BTreeMap<DatabaseVersion, for<'r, 'tx> fn(_: &'r mut DatabaseTransaction<'tx>) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'r>>>,
    decoders: ModuleRegistry<Decoder, DecodingMode>,
) -> Result<(), Error>
where
    F: Fn(Database) -> Fut,
    Fut: Future<Output = Result<(), Error>>,

Validates the database migrations. decoders need to be passed in as an argument since this is module agnostic. First applies all defined migrations to the database then executes the `validate`` function which should confirm the database migrations were successful.