Trait fedimint_testing::btc::BitcoinTest

pub trait BitcoinTest {
    fn mine_blocks(&self, block_num: u64);
    fn send_and_mine_block(
        &self,
        address: &Address,
        amount: Amount
    ) -> (TxOutProof, Transaction);
    fn get_new_address(&self) -> Address;
    fn mine_block_and_get_received(&self, address: &Address) -> Amount;
}

Required Methods

fn mine_blocks(&self, block_num: u64)

Mines a given number of blocks

fn send_and_mine_block(
&self,
address: &Address,
amount: Amount
) -> (TxOutProof, Transaction)

Send some bitcoin to an address then mine a block to confirm it. Returns the proof that the transaction occurred.

fn get_new_address(&self) -> Address

Returns a new address.

fn mine_block_and_get_received(&self, address: &Address) -> Amount

Mine a block to include any pending transactions then get the amount received to an address