Module fedimint_testing::db

Constants

BYTE_8

BYTE_20

BYTE_32

STRING_64

Functions

copy_directory

Helper function that recursively copies all of the contents from atomic_broadcast to dst.

get_project_root

Get the project root (relative to closest Cargo.lock file)

open_temp_db_and_copy 🔒

Open a temporary database located at temp_path and copy the contents from the folder src_dir to the temporary database’s path.

prepare_db_migration_snapshot

Creates the database backup directory by appending the snapshot_name to db/migrations. Then this function will execute the provided prepare_fn which is expected to populate the database with the appropriate data for testing a migration. If the snapshot directory already exists, this function will do nothing.

validate_migrations

Iterates over all of the databases supplied in the database backup directory. First, a temporary database will be created and the contents will be populated from the database backup directory. Next, this function will execute the provided validate closure. The validate closure is expected to do any validation necessary on the temporary database, such as applying the appropriate database migrations and then reading all of the data to verify the migrations were successful.