fedimint_testing::db

Function validate_migrations_core_client

pub async fn validate_migrations_core_client<F, Fut>(
    db_prefix: &str,
    validate: F,
) -> Result<(), Error>
where
    F: Fn(Database) -> Fut,
    Fut: Future<Output = Result<(), Error>>,

Validates the database migrations for the core client. First applies all database migrations to the core client. Then calls the validate function, including the new active_states and inactive_states, and is expected to confirm the database migrations were successful.